all: commit first version of Android client
Barebones of Upspin Android client.
Not adding all the Android Studio and Gradle junk that comes with it.
One will need to import the source into a new Android Studio project and
re-create the project settings and dependencies. We may want to check in
these files later for convenience, but they're not strictly necessary.

Change-Id: Id3b1d75a1155de6c95781222700dd73494411028